Config methods are now abstracted from the main process
